+++ Example/ExampleTests/NSHashTableTests.m
@@ -63,7 +63,7 @@ SHTestsMutableFastEnumerationBlocks
 }
 
 -(void)testMap;{
-  __block NSUInteger skipOne = 0;
+  __block NSInteger skipOne = 0;
   self.matching = [self.subject SH_map:^id(id obj) {
     skipOne += 1;
     if(skipOne == 1)
@@ -98,7 +98,7 @@ SHTestsMutableFastEnumerationBlocks
 }
 
 -(void)testFind;{
-  __block NSUInteger counter = 0;
+  __block NSInteger counter = 0;
   
   id value = [self.subject SH_find:^BOOL(id obj) {
     counter +=1;
@@ -106,13 +106,13 @@ SHTestsMutableFastEnumerationBlocks
   }];
 
   
-  STAssertEquals(self.subject.count, counter, nil);
+  STAssertEquals(self.subject.count, (NSUInteger)counter, nil);
   STAssertTrue([self.subject containsObject:value], nil);
   
 }
 
 -(void)testFindAll;{
-  __block NSUInteger counter = 0;
+  __block NSInteger counter = 0;
   
   self.matching = [self.subject SH_findAll:^BOOL(id obj) {
     counter +=1;
@@ -128,7 +128,7 @@ SHTestsMutableFastEnumerationBlocks
 }
 
 -(void)testReject;{
-  __block NSUInteger counter = 0;
+  __block NSInteger counter = 0;
   
   self.matching = [self.subject SH_reject:^BOOL(id obj) {
     counter +=1;
@@ -344,7 +344,7 @@ SHTestsMutableFastEnumerationBlocks
 
 #pragma mark - <SHTestsMutableFastEnumerationBlocks>
 -(void)testModifyMap; {
-  __block NSUInteger counter = 0;
+  __block NSInteger counter = 0;
   self.matching = self.subject.mutableCopy;
   [self.matching SH_modifyMap:^id(id obj) {
     counter +=1;
@@ -355,9 +355,9 @@ SHTestsMutableFastEnumerationBlocks
   }];
   
 
-  NSUInteger expectedCount = 1;
+  NSInteger expectedCount = 1;
   STAssertTrue(self.matching.count < self.subject.count, nil);
-  STAssertEquals(self.matching.count, expectedCount, nil);
+  STAssertEquals(self.matching.count, (NSUInteger)expectedCount, nil);
   
   for (id obj in self.matching) {
     STAssertTrue([self.subject containsObject:obj], nil);
@@ -368,7 +368,7 @@ SHTestsMutableFastEnumerationBlocks
 }
 
 -(void)testModifyFindAll; {
-  __block NSUInteger counter = 0;
+  __block NSInteger counter = 0;
   self.matching = self.subject.mutableCopy;
   [self.matching SH_modifyFindAll:^BOOL(id obj) {
     counter +=1;
@@ -379,9 +379,9 @@ SHTestsMutableFastEnumerationBlocks
   }];
   
   
-  NSUInteger expectedCount = 1;
+  NSInteger expectedCount = 1;
   STAssertTrue(self.matching.count < self.subject.count, nil);
-  STAssertEquals(self.matching.count, expectedCount, nil);
+  STAssertEquals(self.matching.count, (NSUInteger)expectedCount, nil);
   
   for (id obj in self.matching) {
     STAssertTrue([self.subject containsObject:obj], nil);
@@ -393,7 +393,7 @@ SHTestsMutableFastEnumerationBlocks
 
 -(void)testModifyReject; {
   
-  __block NSUInteger counter = 0;
+  __block NSInteger counter = 0;
   self.matching = self.subject.mutableCopy;
   [self.matching SH_modifyReject:^BOOL(id obj) {
     counter +=1;
@@ -404,7 +404,7 @@ SHTestsMutableFastEnumerationBlocks
   }];
   
   
-  NSUInteger expectedCount = 1;
+  NSInteger expectedCount = 1;
   STAssertTrue(self.matching.count < self.subject.count, nil);
   STAssertEquals(self.matching.count, self.subject.count-expectedCount, nil);
