+++ Example/ExampleTests/NSOrderedSetTests.m
@@ -89,7 +89,7 @@ SHTestsMutableFastEnumerationOrdered>
 }
 
 -(void)testFind;{
-  NSUInteger index = self.subject.count/2;
+  NSInteger index = self.subject.count/2;
   
   id value = [self.subject SH_find:^BOOL(id obj) {
     return [self.subject indexOfObject:obj] == index;
@@ -253,7 +253,7 @@ SHTestsMutableFastEnumerationOrdered>
   STAssertTrue([self.subject.SH_toDictionary isKindOfClass:[NSDictionary class]], nil);
   STAssertTrue(self.subject.SH_toDictionary.count > 0, nil);
   
-  [self.subject SH_eachWithIndex:^(id obj, NSUInteger index) {
+  [self.subject SH_eachWithIndex:^(id obj, NSInteger index) {
     STAssertEqualObjects(self.subject[index], [self.subject.SH_toDictionary objectForKey:@(index)], nil);
   }];
   
@@ -320,7 +320,7 @@ SHTestsMutableFastEnumerationOrdered>
 #pragma mark - <SHTestsFastEnumerationOrderedBlocks>
 -(void)testEachWithIndex;{
   
-  [self.subject SH_eachWithIndex:^(id obj, NSUInteger index) {
+  [self.subject SH_eachWithIndex:^(id obj, NSInteger index) {
     [self.matching addObject:@(index)];
   }];
   
@@ -376,7 +376,7 @@ SHTestsMutableFastEnumerationOrdered>
 
 #pragma mark - <SHTestsMutableFastEnumerationBlocks>
 -(void)testModifyMap; {
-  __block NSUInteger counter = 0;
+  __block NSInteger counter = 0;
   self.matching = self.subject.mutableCopy;
   [self.matching SH_modifyMap:^id(id obj) {
     counter +=1;
@@ -387,9 +387,9 @@ SHTestsMutableFastEnumerationOrdered>
   }];
   
   
-  NSUInteger expectedCount = 1;
+  NSInteger expectedCount = 1;
   STAssertTrue(self.matching.count < self.subject.count, nil);
-  STAssertEquals(self.matching.count, expectedCount, nil);
+  STAssertEquals(self.matching.count, (NSUInteger)expectedCount, nil);
 
   STAssertEqualObjects(self.matching[0], self.subject[0], nil);
   
@@ -498,7 +498,7 @@ SHTestsMutableFastEnumerationOrdered>
   STAssertTrue([theMapTable isKindOfClass:[NSMapTable class]], nil);
   STAssertTrue(theMapTable.count > 0, nil);
   
-  [self.subject SH_eachWithIndex:^(id obj, NSUInteger index) {
+  [self.subject SH_eachWithIndex:^(id obj, NSInteger index) {
     STAssertEqualObjects(self.subject[index], [theMapTable objectForKey:@(index)], nil);
   }];
 }
